// src/app/@core/interfaces/auction/transferentes.model.ts
export interface TransferentesInterface {
    id?: number;
    name?: string;
    keyCode?: string;
    userCreation: string;
    creationDate?: Date;
    modificationUser: string;
    modificationDate?: Date;
    type?: string;
    version?: number;
    status?:string;
    operationStartDate?: Date;
    operationEndDate?: Date;
    sedativePrincipal?: string;
    commissionedObject?: string;
    sector?: string;
    formalization?: string;
    formalizationDate?: Date;
    entity?: string;
    agreementModification?: string;
    agreementDate?: Date;
    goodGuy?: string;
    guardGuardWell?: string;
    destinyWelll?: string;
    daysAdminGood?: string;
    goodReceptionPlace?: string;
}